server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
fonttools==4.60.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.7
more-itertools==10.8.0
numpy==2.3.4
openai==1.106.1
packaging==25.0
pillow==12.0.0
premailer==3.10.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yagmail==0.15.293
